feat: Refactor utils.go
